Fix benchmarks + add profile to build benchmarks.jar + add GitHub workflow for benchmarks.jar (Closes #885)
