+++ drasyl-performance-tests/src/main/java/org/drasyl/crypto/SignBenchmark.java
