+++ drasyl-performance-tests/src/main/java/org/drasyl/handler/remote/protocol/ArmedProtocolMessageBenchmark.java
