+++ drasyl-performance-tests/src/main/java/org/drasyl/node/handler/serialization/StringSerializerBenchmark.java
