+++ drasyl-performance-tests/src/main/java/org/drasyl/util/NumberUtilBenchmark.java
