+++ drasyl-performance-tests/src/main/java/org/drasyl/util/logging/LoggerBenchmark.java
